+++ api/src/types/collection.ts
@@ -8,7 +8,7 @@ export type Collection = {
 		collection: string;
 		note: string | null;
 		hidden: boolean;
-		single: boolean;
+		singleton: boolean;
 		icon: string | null;
 		translation: Record<string, string>;
 	} | null;
